# gh-129005: Move bytearray to use bytes as a buffer