--- a/Misc/NEWS.d/next/Core_and_Builtins/2025-02-25-20-31-36.gh-issue-129005.0KBps9.rst
+++ b/Misc/NEWS.d/next/Core_and_Builtins/2025-02-25-20-31-36.gh-issue-129005.0KBps9.rst
@@ -1,0 +1,1 @@
+Move :class:`bytearray` to use a buffer convertable to :class:`bytes` without copying.