@cmaloney 2025-03-12
Closing while discussion is ongoing, if decide to ship, will make a new github issue this will connect to.